+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 1.0.11
+
 * Domain that starts and ends with '-' are not valid anymore in ENS.
 * Bugfix Resolution#resolve on ENS domain when resolver has no address record
 * Resolution#isValidHash method - checks wheather a domain name matches the given hash from the blockchain